+++ NEWS.md
@@ -1,2 +1,7 @@
-# FSA 0.0.1 ongoing
+# RFishBC 0.0.2 ongoing
+* `digitizeRadii()`: Modified. Added `snap2Transect=` to address [#1](https://github.com/droglenc/RFishBC/issues/1). Moved `file.choose()` for `fname=` out of argument list and into the main function code. Fixed poor directions about the use of the escape key to terminate `locator()`.
+* `iSnap2Transect()`: Added. Used in `digitizeRadii()`.
+* `showDigitizedData()`: Modified. Moved `file.choose()` for `fname=` out of argument list and into the main function code.
+
+# RFishBC 0.0.1 16-Apr-18
 * First version. Everything is new.